backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
openai
google-generativeai
PyMuPDF
pandas
openpyxl
python-multipart
reportlab
python-dotenv
numpy
mixpanel
supabase
